requirements.txt
gym==0.14.0
numpy==1.18.4
scipy==1.1.0
lxml==4.4.1
pyprind==2.11.2
nose2==0.8.0
six==1.11.0
path.py
joblib==0.10.3
python-dateutil==2.7.3
cached_property
cloudpickle==1.2.0
pyglet==1.3.2
matplotlib==3.1.0
imutils==0.5.1
numpydoc
ray==0.7.3
opencv-python
dill
lz4
setproctitle
psutil
opencv-python
boto3==1.4.8
redis~=2.10.6
pandas==0.24.2
plotly==2.4.0
pillow==7.1.2
scikit-image==0.17.2
wandb==0.9.1
torch==1.5.0
ray[rllib]
torchvision==0.6.0
